+++ src/host_net/tcp_host.rs
@@ -23,7 +23,7 @@ pub async fn wait_for_remote_message(
 }
 
 /// Send config change received form the GUI to the remote hardware over `stream`[TcpStream]
-pub async fn send_config_change(
+pub async fn send_config_message(
     mut stream: TcpStream,
     config_change_message: &HardwareConfigMessage,
 ) -> anyhow::Result<()> {
